GITISSUE-112 refactor: consolidate validation and mocks into pkg/utils
- Moved validation functions from `pkg/validation` to `pkg/utils` to centralize utility functions.
- Updated all handlers to use `utils.ValidateID` and `utils.ValidatePagination` instead of the deprecated `validation` package.
- Relocated `mocks` package from `internal/mocks` to `pkg/mocks` for better organization and accessibility across the project.
- Updated import paths in handlers, tests, and smoke tests to reflect the new package structure.
- Renamed `pkg/validation/validation.go` to `pkg/utils/validation.go` to align with the new structure.
