+++ pkg/handlers/anime_handlers.go
@@ -21,7 +21,8 @@ import (
 	"myanimeapi/internal/db"
 	"myanimeapi/pkg/middleware"
 	"myanimeapi/pkg/models"
-	"myanimeapi/pkg/validation"
+
+	"myanimeapi/pkg/utils"
 
 	"github.com/gorilla/mux"
 )
@@ -74,7 +75,7 @@ func (h *AnimeHandler) GetAnimeHandler(w http.ResponseWriter, r *http.Request) {
 	idStr := vars["id"]
 
 	// Validate ID
-	id, err := validation.ValidateID(idStr)
+	id, err := utils.ValidateID(idStr)
 	if err != nil {
 		log.Printf("Invalid ID format: %v", err)
 		http.Error(w, err.Error(), http.StatusBadRequest)
@@ -175,7 +176,7 @@ func (h *AnimeHandler) GetPaginatedReviewsForAnimeHandler(w http.ResponseWriter,
 	idStr := vars["id"]
 
 	// Validate ID
-	id, err := validation.ValidateID(idStr)
+	id, err := utils.ValidateID(idStr)
 	if err != nil {
 		log.Printf("Invalid ID format: %v", err)
 		http.Error(w, err.Error(), http.StatusBadRequest)
@@ -185,7 +186,7 @@ func (h *AnimeHandler) GetPaginatedReviewsForAnimeHandler(w http.ResponseWriter,
 	// Validate pagination
 	pageStr := r.URL.Query().Get("page")
 	limitStr := r.URL.Query().Get("limit")
-	page, limit, err := validation.ValidatePagination(pageStr, limitStr, 1, 10)
+	page, limit, err := utils.ValidatePagination(pageStr, limitStr, 1, 10)
 	if err != nil {
 		log.Printf("Invalid pagination parameters: %v", err)
 		http.Error(w, err.Error(), http.StatusBadRequest)
@@ -310,7 +311,7 @@ func (h *AnimeHandler) UpdateAnimeHandler(w http.ResponseWriter, r *http.Request
 	idStr := vars["id"]
 
 	// Validate ID
-	id, err := validation.ValidateID(idStr)
+	id, err := utils.ValidateID(idStr)
 	if err != nil {
 		log.Printf("Invalid ID format: %v", err)
 		http.Error(w, err.Error(), http.StatusBadRequest)
@@ -370,7 +371,7 @@ func (h *AnimeHandler) DeleteAnimeHandler(w http.ResponseWriter, r *http.Request
 	idStr := vars["id"]
 
 	// Validate ID
-	id, err := validation.ValidateID(idStr)
+	id, err := utils.ValidateID(idStr)
 	if err != nil {
 		log.Printf("Invalid ID format: %v", err)
 		http.Error(w, err.Error(), http.StatusBadRequest)
