+++ pkg/handlers/anime_handlers_test.go
@@ -9,9 +9,9 @@ import (
 	"context"
 	"encoding/json"
 	"errors"
-	"myanimeapi/internal/mocks"
 	"myanimeapi/pkg/handlers"
 	"myanimeapi/pkg/middleware"
+	"myanimeapi/pkg/mocks"
 	"myanimeapi/pkg/models"
 	"net/http"
 	"net/http/httptest"
