+++ pkg/utils/validation.go
@@ -1,7 +1,7 @@
-// pkg/validation/validation.go
+// pkg/utils/validation.go
 // Package validation provides utility functions for validating and parsing input data.
 // It includes functions for validating IDs and pagination parameters.
-package validation
+package utils
 
 import (
 	"errors"
